scripts/flight_fetch_serpapi/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-cloud-storage>=2.10.0